+++ processor/src/test/resources/expected/Arez_BasicObservableModel.java
@@ -86,6 +86,7 @@ public final class Arez_BasicObservableModel extends BasicObservableModel implem
   public void setTime(final long time) {
     Guards.invariant( () -> !this.$$arez$$_disposed, () -> "Method invoked on invalid component '" + $$arez$$_name() + "'" );
     if ( time != super.getTime() ) {
+      this.$$arez$$_time.preReportChanged();
       super.setTime(time);
       this.$$arez$$_time.reportChanged();
     }