+++ processor/src/test/resources/expected/Arez_SingletonModel.java
@@ -106,6 +106,7 @@ public final class Arez_SingletonModel extends SingletonModel implements Disposa
   public void setTime(final long time) {
     Guards.invariant( () -> !this.$$arez$$_disposed, () -> "Method invoked on invalid component '" + $$arez$$_name() + "'" );
     if ( time != super.getTime() ) {
+      this.$$arez$$_time.preReportChanged();
       super.setTime(time);
       this.$$arez$$_time.reportChanged();
     }