+++ processor/src/test/resources/expected/com/example/deprecated/Arez_DeprecatedObservableModel2.java
@@ -90,6 +90,7 @@ public final class Arez_DeprecatedObservableModel2 extends DeprecatedObservableM
   void setTime(final long time) {
     Guards.invariant( () -> !this.$$arez$$_disposed, () -> "Method invoked on invalid component '" + $$arez$$_name() + "'" );
     if ( time != super.getTime() ) {
+      this.$$arez$$_time.preReportChanged();
       super.setTime(time);
       this.$$arez$$_time.reportChanged();
     }