+++ processor/src/test/resources/expected/com/example/repository/Arez_RepositoryWithExplicitId.java
@@ -95,6 +95,7 @@ final class Arez_RepositoryWithExplicitId extends RepositoryWithExplicitId imple
   public void setName(@Nonnull final String name) {
     Guards.invariant( () -> !this.$$arez$$_disposed, () -> "Method invoked on invalid component '" + $$arez$$_name() + "'" );
     if ( !Objects.equals(name, super.getName()) ) {
+      this.$$arez$$_name.preReportChanged();
       super.setName(name);
       this.$$arez$$_name.reportChanged();
     }