+++ processor/src/test/resources/expected/com/example/type_access_levels/Arez_ReduceAccessLevelModel.java
@@ -88,6 +88,7 @@ final class Arez_ReduceAccessLevelModel extends ReduceAccessLevelModel implement
   public void setTime(final long time) {
     Guards.invariant( () -> !this.$$arez$$_disposed, () -> "Method invoked on invalid component '" + $$arez$$_name() + "'" );
     if ( time != super.getTime() ) {
+      this.$$arez$$_time.preReportChanged();
       super.setTime(time);
       this.$$arez$$_time.reportChanged();
     }